chore: move shadcn/ui files to `shadcn-ui/`
