+++ app/shadcn-ui/components/$collapsible.tsx
