+++ app/shadcn-ui/components/avatar.tsx
@@ -1,7 +1,7 @@
 import * as AvatarPrimitive from "@radix-ui/react-avatar"
 import type * as React from "react"
 
-import { cn } from "@/lib/utils"
+import { cn } from "@/shadcn-ui/lib/utils"
 
 function Avatar({
   className,
