+++ app/shadcn-ui/components/breadcrumb.tsx
@@ -2,7 +2,7 @@ import { Slot } from "@radix-ui/react-slot"
 import { ChevronRight, MoreHorizontal } from "lucide-react"
 import type * as React from "react"
 
-import { cn } from "@/lib/utils"
+import { cn } from "@/shadcn-ui/lib/utils"
 
 function Breadcrumb({ ...props }: React.ComponentProps<"nav">) {
   return <nav aria-label="breadcrumb" data-slot="breadcrumb" {...props} />
