+++ app/shadcn-ui/components/tooltip.tsx
@@ -1,7 +1,7 @@
 import * as TooltipPrimitive from "@radix-ui/react-tooltip"
 import type * as React from "react"
 
-import { cn } from "@/lib/utils"
+import { cn } from "@/shadcn-ui/lib/utils"
 
 function TooltipProvider({
   delayDuration = 0,
